Stage 2.5 (section-align Events): consolidate ViewModels under Models/Events/
Moves the 5 grab-bag VM files plus the events portion of
DashboardViewModel.cs under src/Humans.Web/Models/Events/ with the
matching Humans.Web.Models.Events sub-namespace (folder-mirrors-namespace
convention, same as Models/Camp/, Models/Google/, etc.):

  CampEventViewModels.cs      → Events/BarrioEventViewModels.cs
  GuideAdminViewModels.cs     → Events/EventsAdminViewModels.cs
  GuideApiModels.cs           → Events/EventsApiModels.cs
  IndividualEventViewModels.cs → Events/IndividualEventViewModels.cs
  ModerationViewModels.cs     → Events/EventsModerationViewModels.cs

DashboardViewModel.cs was mixed — the home-dashboard VM
(DashboardViewModel + MyAttendeeRowVm, used by HomeController) stays
in place; the events-dashboard portion (GuideDashboardViewModel + the
three coverage row types) was extracted to a new
Events/EventsDashboardViewModel.cs, with GuideDashboardViewModel
renamed to EventsDashboardViewModel.

_ViewImports.cshtml gains @using Humans.Web.Models.Events.
Controllers gain using Humans.Web.Models.Events;.
Razor views' Humans.Web.Models.<MovedType> FQNs rewritten to
Humans.Web.Models.Events.<MovedType>.

Internal class names (GuideSettingsViewModel, GuideVenueListViewModel,
GuideEventApiDto, etc.) keep their old names — file/folder/namespace
alignment is the Stage 2.5 deliverable; symbol-level polish is Stage 3.
