+++ src/Humans.Web/Controllers/GuideAdminController.cs
@@ -3,6 +3,7 @@ using Humans.Domain.Constants;
 using Humans.Domain.Entities;
 using Humans.Web.Filters;
 using Humans.Web.Models;
+using Humans.Web.Models.Events;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
