+++ src/Humans.Web/Views/CampEvents/Index.cshtml
@@ -1,4 +1,4 @@
-@model Humans.Web.Models.CampEventsTabViewModel
+@model Humans.Web.Models.Events.CampEventsTabViewModel
 @{
     ViewData["Title"] = $"Events — {Model.CampName}";
 }
