+++ src/Humans.Web/Views/EventGuide/Browse.cshtml
@@ -1,4 +1,4 @@
-@model Humans.Web.Models.BrowseViewModel
+@model Humans.Web.Models.Events.BrowseViewModel
 @{
     ViewData["Title"] = "Browse Events";
     var hasFilters = Model.FilterDays.Count > 0 || Model.FilterCategoryId.HasValue || Model.FilterVenueId.HasValue
