+++ src/Humans.Web/Views/EventGuide/IndividualEventForm.cshtml
@@ -1,4 +1,4 @@
-@model Humans.Web.Models.IndividualEventFormViewModel
+@model Humans.Web.Models.Events.IndividualEventFormViewModel
 @{
     var isEdit = Model.Id.HasValue;
     ViewData["Title"] = isEdit ? "Edit Event" : "Submit an Event";
