+++ src/Humans.Web/Views/EventGuide/MySubmissions.cshtml
@@ -1,4 +1,4 @@
-@model Humans.Web.Models.MySubmissionsViewModel
+@model Humans.Web.Models.Events.MySubmissionsViewModel
 @{
     ViewData["Title"] = "My Event Submissions";
 }
