+++ src/Humans.Web/Views/EventGuide/Schedule.cshtml
@@ -1,4 +1,4 @@
-@model Humans.Web.Models.ScheduleViewModel
+@model Humans.Web.Models.Events.ScheduleViewModel
 @{
     ViewData["Title"] = "My Schedule";
 }
