+++ src/Humans.Web/Views/EventGuideDashboard/Index.cshtml
@@ -1,4 +1,4 @@
-@model Humans.Web.Models.GuideDashboardViewModel
+@model Humans.Web.Models.Events.EventsDashboardViewModel
 @{
     ViewData["Title"] = "Event Guide Dashboard";
 }
