+++ src/Humans.Web/Views/GuideAdmin/GuideCategories.cshtml
@@ -1,4 +1,4 @@
-@model Humans.Web.Models.EventCategoryListViewModel
+@model Humans.Web.Models.Events.EventCategoryListViewModel
 @{
     ViewData["Title"] = "Event Categories";
 }
