+++ src/Humans.Web/Views/GuideAdmin/GuideCategoryForm.cshtml
@@ -1,4 +1,4 @@
-@model Humans.Web.Models.EventCategoryFormViewModel
+@model Humans.Web.Models.Events.EventCategoryFormViewModel
 @{
     var isEdit = Model.Id.HasValue;
     ViewData["Title"] = isEdit ? "Edit Category" : "New Category";
