+++ src/Humans.Web/Views/GuideAdmin/GuideSettings.cshtml
@@ -1,4 +1,4 @@
-@model Humans.Web.Models.GuideSettingsViewModel
+@model Humans.Web.Models.Events.GuideSettingsViewModel
 @{
     ViewData["Title"] = "Guide Settings";
     var isEdit = Model.Id.HasValue;
