+++ src/Humans.Web/Views/GuideAdmin/GuideVenueForm.cshtml
@@ -1,4 +1,4 @@
-@model Humans.Web.Models.GuideVenueFormViewModel
+@model Humans.Web.Models.Events.GuideVenueFormViewModel
 @{
     var isEdit = Model.Id.HasValue;
     ViewData["Title"] = isEdit ? "Edit Venue" : "New Venue";
