+++ src/Humans.Web/Views/GuideAdmin/GuideVenues.cshtml
@@ -1,4 +1,4 @@
-@model Humans.Web.Models.GuideVenueListViewModel
+@model Humans.Web.Models.Events.GuideVenueListViewModel
 @{
     ViewData["Title"] = "Communal Venues";
 }
