+++ src/Humans.Web/Views/Moderation/Index.cshtml
@@ -1,4 +1,4 @@
-@model Humans.Web.Models.ModerationQueueViewModel
+@model Humans.Web.Models.Events.ModerationQueueViewModel
 @using Humans.Domain.Enums
 @{
     ViewData["Title"] = "Moderate Events";
