+++ src/Humans.Web/Views/_ViewImports.cshtml
@@ -3,6 +3,7 @@
 @using Humans.Web.Extensions.Sections
 @using Humans.Web.Helpers
 @using Humans.Web.Models
+@using Humans.Web.Models.Events
 @using Humans.Web.Models.Google
 @using Humans.Web.ViewComponents
 @using Humans.Domain.Entities
